Note/backendapi/requirements.txt
flask
flask-restful
google-cloud-firestore
gunicorn
marshmallow
wheel